+++ nifti2bids/__init__.py
@@ -1,1 +1,20 @@
-__version__ = "0.0.9"
+"""
+Post-hoc BIDS conversion toolkit for NIfTI datasets without original DICOMs.
+----------------------------------------------------------------------------
+Documentation can be found at https://nifti2bids.readthedocs.io.
+
+Submodules
+----------
+bids -- Operations related to initializing and creating BIDs compliant files
+
+io -- Generic operations related to loading NIfTI data
+
+logging -- Set up a logger using ``RichHandler`` as the default handler if a root or
+module specific handler is not available
+
+metadata -- Operations related to extracting metadata information from NIfTI images
+
+simulate -- Simulate a basic NIfTI image for testing purposes
+"""
+
+__version__ = "0.1.0"